+++ llmexer/exceptions.py
@@ -3,3 +3,7 @@
 
 class LLMExerException(Exception):
     """Generic exception for LLMExerException."""
+
+
+class ExperimentAlreadyExistsException(LLMExerException):
+    """Raised when an experiment with the given ID already exists."""