+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.1.2 (2026-03-30)
+------------------
+
+* Modify command that crates experiments, so it would accept custom ID
+* Add command to `list` experiments
+* Add tests
+
 0.1.1 (2026-03-30)
 ------------------